src: make each plot use its own args
