+++ src/env.rs
@@ -9,7 +9,7 @@ impl Env {
     pub const SYS_NAME: &'static str = "sc2-exp";
 
     pub fn proj_root() -> PathBuf {
-        env::current_dir().expect("sc2-exp(env): failed to get current directory")
+        env!("CARGO_MANIFEST_DIR").into()
     }
 
     pub fn results_root() -> PathBuf {
