+++ src/main.rs
@@ -1,7 +1,12 @@
-use crate::experiment::{AvailableExperiments, Exp, ExpRunArgs};
-use crate::plot::Plot;
-use clap::{Parser, Subcommand};
+use crate::{
+    args::{ExpCommand, ImagePullSubCommand, ScaleOutSubCommand, StartUpSubCommand},
+    experiment::{AvailableExperiments, Exp},
+    plot::Plot,
+};
+use clap::Parser;
 
+pub mod args;
+pub mod baselines;
 pub mod containerd;
 pub mod cri;
 pub mod deploy;
@@ -18,33 +23,6 @@ struct Cli {
 
     #[arg(short, long, global = true)]
     debug: bool,
-}
-
-#[derive(Debug, Subcommand)]
-enum ExpSubCommand {
-    /// Run
-    Run(ExpRunArgs),
-    /// Plot
-    Plot {},
-}
-
-#[derive(Debug, Subcommand)]
-enum ExpCommand {
-    /// Experiment comparing different image pulling techniques
-    ImagePull {
-        #[command(subcommand)]
-        exp_sub_command: ExpSubCommand,
-    },
-    /// Evaluate the start-up latency
-    StartUp {
-        #[command(subcommand)]
-        exp_sub_command: ExpSubCommand,
-    },
-    /// Evaluate scale-out latency
-    ScaleOut {
-        #[command(subcommand)]
-        exp_sub_command: ExpSubCommand,
-    },
 }
 
 fn main() {
@@ -67,30 +45,30 @@ fn main() {
         ExpCommand::ImagePull {
             exp_sub_command: eval_sub_command,
         } => match eval_sub_command {
-            ExpSubCommand::Run(run_args) => {
-                Exp::run(&AvailableExperiments::ImagePull, run_args);
+            ImagePullSubCommand::Run(run_args) => {
+                Exp::run_image_pull(run_args);
             }
-            ExpSubCommand::Plot {} => {
+            ImagePullSubCommand::Plot {} => {
                 Plot::plot(&AvailableExperiments::ImagePull);
             }
         },
         ExpCommand::ScaleOut {
             exp_sub_command: eval_sub_command,
         } => match eval_sub_command {
-            ExpSubCommand::Run(run_args) => {
-                Exp::run(&AvailableExperiments::ScaleOut, run_args);
+            ScaleOutSubCommand::Run(run_args) => {
+                Exp::run_scale_out(run_args);
             }
-            ExpSubCommand::Plot {} => {
+            ScaleOutSubCommand::Plot {} => {
                 Plot::plot(&AvailableExperiments::ScaleOut);
             }
         },
         ExpCommand::StartUp {
             exp_sub_command: eval_sub_command,
         } => match eval_sub_command {
-            ExpSubCommand::Run(run_args) => {
-                Exp::run(&AvailableExperiments::StartUp, run_args);
+            StartUpSubCommand::Run(run_args) => {
+                Exp::run_start_up(run_args);
             }
-            ExpSubCommand::Plot {} => {
+            StartUpSubCommand::Plot {} => {
                 Plot::plot(&AvailableExperiments::StartUp);
             }
         },
